+++ processors/baggage/baggagetrace/doc.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Package baggage implements the baggage span processor which duplicates
+// package baggagetrace implements the baggage span processor which duplicates
 // onto a span the attributes found in Baggage in the parent context at
 // the moment the span is started.
-package baggage // import "go.opentelemetry.io/contrib/processors/baggage"
+package baggagetrace // import "go.opentelemetry.io/contrib/processors/baggage/baggagetrace"
